Version 1.0

+++ src/main/java/es/santander/ascender/ejerc006/service/AulaService.java
@@ -6,13 +6,17 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import es.santander.ascender.ejerc006.model.Aula;
+import es.santander.ascender.ejerc006.model.Mesa;
 import es.santander.ascender.ejerc006.repository.AulaRepository;
+import es.santander.ascender.ejerc006.repository.MesaRepository;
 
 @Service
 public class AulaService {
 
     @Autowired
     private AulaRepository aulaRepository;
+     @Autowired
+    private MesaRepository mesaRepository;
 
     // Crear un Aula
     public Aula create(Aula aula) {
@@ -46,6 +50,11 @@ public class AulaService {
 
     // Eliminar un Aula por ID
     public void delete(Long id) {
+        // Verifico si hay mesas asociadas al aula, si hay no se elimina el aula
+        List<Mesa> mesas = mesaRepository.findByAulaId(id);
+        if (!mesas.isEmpty()) {
+            throw new CrudSecurityException("No se puede eliminar el aula porque tiene mesas asociadas", CRUDOperation.DELETE, id);
+        }
         aulaRepository.deleteById(id);
     }
 }
